api/requirements.txt
APScheduler==3.9.1.post1
arrow==1.2.3
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
blessed==1.19.1
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
Django==4.1.3
django-apscheduler==0.6.2
django-picklefield==3.1
django-rest-framework==0.1.0
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pip==22.3.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
requests==2.28.1
service-identity==21.1.0
setuptools==58.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
wcwidth==0.2.5
websockets==10.4
zope.interface==5.5.2
